python/requirements.txt
pyserial
influxdb
pyyaml
